+++ upstream/solr4/local/91-healthcheck.sh
@@ -19,12 +19,13 @@ export PORT
 echo "#!/bin/bash" >${SOLR_INSTALL_HOME}/healthcheck.sh
 if [ $ALFRESCO_SSL != none ]
 then
-    if [ ! -f "${SOLR_INSTALL_HOME}/keystore/browser.p12" ]
+    # for custom certificates, replace browser.pem with certificates able to talk to solr
+    # this only works because images use the same keystore for alfresco and solr
+    if [ ! -f "${SOLR_INSTALL_HOME}/keystore/browser.pem" ]
     then
         keytool -importkeystore -srckeystore "${SOLR_INSTALL_HOME}/keystore/${SSL_KEY_STORE}" -srcstorepass ${SSL_KEY_STORE_PASSWORD} -srcstoretype JCEKS -srcalias ssl.repo -destkeystore  "${SOLR_INSTALL_HOME}/keystore/browser.p12" -deststoretype pkcs12 -destalias ssl.repo -deststorepass alfresco -destkeypass alfresco
         openssl pkcs12 -in "${SOLR_INSTALL_HOME}/keystore/browser.p12" -out "${SOLR_INSTALL_HOME}/keystore/browser.pem" -nodes -passin pass:alfresco
     fi
-    # for custom certificates, replace browser.pem with certificates able to talk to solr
     echo "status=\$(curl -f -k -L -w %{http_code} -s -E ${SOLR_INSTALL_HOME}/keystore/browser.pem -o /dev/null https://localhost:${PORT}/solr4)" >>${SOLR_INSTALL_HOME}/healthcheck.sh
 else
         echo "status=\$(curl -f -L -w %{http_code} -s -o /dev/null http://localhost:${PORT}/solr4)" >>${SOLR_INSTALL_HOME}/healthcheck.sh