+++ upstream/solr6/local/91-healthcheck.sh
@@ -20,12 +20,13 @@ export PORT
 echo "#!/bin/bash" >${SOLR_INSTALL_HOME}/healthcheck.sh
 if [ $ALFRESCO_SSL != none ]
 then
-    if [ ! -f "${SOLR_DIR_ROOT}/keystore/browser.p12" ]
+    # for custom certificates, replace browser.pem with certificates able to talk to solr
+    # this only works because images use the same keystore for alfresco and solr
+    if [ ! -f "${SOLR_DIR_ROOT}/keystore/browser.pem" ]
     then
-	keytool -importkeystore -srckeystore "${SOLR_DIR_ROOT}/keystore/${SSL_KEY_STORE}" -srcstorepass ${SSL_KEY_STORE_PASSWORD} -srcstoretype JCEKS -srcalias ssl.repo -destkeystore  "${SOLR_DIR_ROOT}/keystore/browser.p12" -deststoretype pkcs12 -destalias ssl.repo -deststorepass alfresco -destkeypass alfresco
-	openssl pkcs12 -in "${SOLR_DIR_ROOT}/keystore/browser.p12" -out "${SOLR_DIR_ROOT}/keystore/browser.pem" -nodes -passin pass:alfresco
+	    keytool -importkeystore -srckeystore "${SOLR_DIR_ROOT}/keystore/${SSL_KEY_STORE}" -srcstorepass ${SSL_KEY_STORE_PASSWORD} -srcstoretype JCEKS -srcalias ssl.repo -destkeystore  "${SOLR_DIR_ROOT}/keystore/browser.p12" -deststoretype pkcs12 -destalias ssl.repo -deststorepass alfresco -destkeypass alfresco
+	    openssl pkcs12 -in "${SOLR_DIR_ROOT}/keystore/browser.p12" -out "${SOLR_DIR_ROOT}/keystore/browser.pem" -nodes -passin pass:alfresco
     fi
-    # for custom certificates, replace the browser.pem with certificates able to talk to solr    
     echo "status=\$(curl -f -k -L -w %{http_code} -s -E ${SOLR_DIR_ROOT}/keystore/browser.pem -o /dev/null https://localhost:${PORT}/solr)" >>${SOLR_INSTALL_HOME}/healthcheck.sh
 else
     echo "status=\$(curl -f -L -w %{http_code} -s -o /dev/null http://localhost:${PORT}/solr)" >>${SOLR_INSTALL_HOME}/healthcheck.sh